src/opentts_tensorflow_integration/tensorflow_tts/tensorflow_tts_wrapper.py
import numpy as np
import tensorflow as tf
from ..core import BaseTTS
from ..utils import (
    log_error,
    log_info,
    TTSError,
    ModelNotFoundError,
    VoiceNotFoundError,
    AudioConversionError
)
from ..utils.i18n import _

class TensorFlowTTSWrapper(BaseTTS):
    """
    Wrapper class for TensorFlow TTS integration.
    """

    def __init__(self):
        super().__init__()
        self.model = None
        self.processor = None
        log_info(_("Initialized TensorFlowTTSWrapper"))

    def load_model(self, model_path: str):
        """
        Load TensorFlow TTS model.

        Args:
            model_path (str): Path to the TensorFlow TTS model.
        """
        try:
            import tensorflow_tts
            self.model = tf.saved_model.load(model_path)
            # Assuming the processor is saved alongside the model
            self.processor = tensorflow_tts.processor.load(f"{model_path}_processor")
            log_info(_("Loaded TensorFlow TTS model from {}").format(model_path))
        except ImportError:
            log_error(_("TensorFlow TTS is not installed"))
            raise ImportError(_("TensorFlow TTS is not installed. Please install it with 'pip install TensorFlowTTS'."))
        except Exception as e:
            log_error(_("Failed to load TensorFlow TTS model: {}").format(str(e)))
            raise ModelNotFoundError(_("Failed to load TensorFlow TTS model: {}").format(str(e)))

    def synthesize(self, text: str) -> np.ndarray:
        """
        Convert text to speech using TensorFlow TTS.

        Args:
            text (str): The input text to be converted to speech.

        Returns:
            np.ndarray: The audio waveform as a numpy array.
        """
        if self.model is None or self.processor is None:
            log_error(_("TensorFlow TTS model is not initialized"))
            raise TTSError(_("TensorFlow TTS model is not initialized. Please call load_model() first."))

        try:
            normalized_text = self.normalize_text(text)
            log_info(_("Synthesizing text: '{}'").format(normalized_text))
            input_ids = self.processor.text_to_sequence(normalized_text)
            
            # Assuming the model takes input_ids and returns mel spectrograms
            mel_outputs = self.model.generate_mel(input_ids)
            
            # Convert mel spectrograms to audio waveform
            audio = self.vocoder(mel_outputs)
            
            return audio.numpy()
        except Exception as e:
            log_error(_("Error during speech synthesis: {}").format(str(e)))
            raise TTSError(_("Failed to synthesize speech: {}").format(str(e)))

    @property
    def available_voices(self) -> list:
        """
        Get the list of available voices in TensorFlow TTS.

        Returns:
            list: A list of available voice names.
        """
        # This might need to be implemented differently depending on how
        # TensorFlow TTS handles multiple voices
        log_info(_("Retrieving available voices for TensorFlow TTS"))
        return ["default_voice"]

    def set_voice(self, voice_name: str):
        """
        Set the current voice for speech synthesis in TensorFlow TTS.

        Args:
            voice_name (str): The name of the voice to use.
        """
        try:
            if voice_name not in self.available_voices:
                raise VoiceNotFoundError(_("Voice '{}' is not available in TensorFlow TTS.").format(voice_name))

            # This might need to be implemented differently depending on how
            # TensorFlow TTS handles multiple voices
            self.current_voice = voice_name
            log_info(_("Set current voice to '{}'").format(voice_name))
        except VoiceNotFoundError as e:
            log_error(str(e))
            raise
        except Exception as e:
            log_error(_("Error setting voice: {}").format(str(e)))
            raise TTSError(_("Failed to set voice: {}").format(str(e)))

    def vocoder(self, mel_outputs):
        """
        Convert mel spectrograms to audio waveform.

        Args:
            mel_outputs: Mel spectrogram output from the TTS model.

        Returns:
            Audio waveform.
        """
        # This is a placeholder. You'll need to implement this using the
        # specific vocoder you're using with TensorFlow TTS.
        log_error(_("Vocoder not implemented"))
        raise NotImplementedError(_("Vocoder not implemented. Please implement this method using your chosen vocoder."))